pyproject.toml
[project]
name = "bank_transactions_analyzer"
version = "1.0.0"
description = "Анализатор банковских транзакций"
authors = [
    {name = "Curator", email = "curator@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.4,<9.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
mypy = "^1.14.1"
isort = "^5.13.2"
flake8 = "^7.1.1"
black = "^25.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[tool.black]
line-length = 119
exclude = ".git"

[tool.isort]
profile = "black"
line_length = 119

[tool.mypy]
strict = true
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"